requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
iniconfig==2.3.0
packaging==25.0
piexif==1.1.3
pillow==12.0.0
pluggy==1.6.0
PyAudio==0.2.14
Pygments==2.19.2
pytest==9.0.1
requests==2.32.5
SpeechRecognition==3.14.3
typing_extensions==4.15.0
urllib3==2.5.0
